requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
ta>=0.10.2
python-telegram-bot>=20.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
Pillow>=10.0.0
python-dotenv>=1.0.0
asyncio
